Implement StructuredGrid.concatenate() without VTK
I did not understand the behavior of vtkStructuredGridAppend. It permits non-adjacent grids to be joined. The axis of the join is somehow implicit. I didn't find much documentation or usages of it elsewhere, so I figured implementing concatenate() within pyvista would be a better yet.

The trade-off is that there is a lot more code and test baggage to maintain. Also, now that there is an argument for concatenate, the __add__ behavior is not as easily inferred, so I took that out for now.