add DICOMReader (#2460)
* Add `DICOMReader` and `pancreas` example

* Add `test_dcmreader`

* Change `download` and `test` to DICOM stack

DICOM dataset was altered as is no longer a pancreas scan. Rather
than explicitly name the type of scan, simply refer to it as a
dicom stack, which is what is important, in case the data needs to
change.

* Call `DICOMReader` in `test_dcmreader` for folder

`pyvista/utilities/reader/get_reader` requires a file extension to
work (it does not support folders). Since we would have to specify
explicitly the files to be read from teh folder, we call the reader
directly. However, we still support `get_reader` for reading a single
`.dcm` file.

* Update `download_dicom_stack` to pull from `VTK_DATA_PATH`

`pyvista/vtk-data` PR #5 was accepted. Pull dicom stack from here
instead of temporary local directory.

* Fix `pre-commit` errors

* Incorporate PR #2460 feedback

- Use `_download_and_read` for `dicom_stack`
- Add volume plot in Examples docstrings
- Update `DICOMReader` to accept files or folders
- Remove redundant `dir_path` assignment in `examples.py`
- Add test for valid mesh after loading single *.dcm file
- Add test to `test_examples.py`
- Update `check-jsonschema` to latest version

* cleanup

* Update pyvista/examples/downloads.py

Co-authored-by: MatthewFlamm <39341281+MatthewFlamm@users.noreply.github.com>

* cleanup common methods

* small test cleanup

* remove extension check; reset cache

* remove warning test

* improve coverage

* remove __init__ in DICOMReader

* refactor filename --> path

* final cleanup

* reset directory and filename when setting filename and directory

Co-authored-by: Adam Hendry <adam.hendry@medtronic.com>
Co-authored-by: Alex Kaszynski <akascap@gmail.com>
Co-authored-by: MatthewFlamm <39341281+MatthewFlamm@users.noreply.github.com>