general fixes of invalid links
Some links are invalid or not available,
they need to be removed or replaced.

Some of GitHub links have a different way of
adding anchors. You can read more here:

https://github.com/sphinx-doc/sphinx/issues/9016

Some links seem to be working but fail
on the linkcheck. For those, I've added to
conf.py to not run the linkcheck.