fix(hack): replace `HACK` with memory guard check
`add_volume` rescales `scalars` to 8-bit for colormapping when
plotting. We should not be mutating original data, so replace `HACK`
with explicit check that `scalars` is not the original data and force
a copy.

Also, remove added `kwargs` to method since choosing whether or not to
scale data is likely not needed in most cases (and unrelated to the
original bug that spawned this PR).