fix(add_volume): use memory-efficient rescaling (#3170)
* fix(add_volume): use memory-efficient rescaling

Inefficient `numpy` copies causes memory leak for large volume
data sets. Use a vectorized rescaling instead.

See `pyvista/support` Issue #500:
    - https://github.com/pyvista/pyvista-support/issues/500

* fix(quotes): revert rogue changes to quotes

* fix(mutation): use copy to prevent overwritting

Copy the `volume.active_scalars` so we do not mutate its values when
rescaling.

* fix(hack): replace `HACK` with memory guard check

`add_volume` rescales `scalars` to 8-bit for colormapping when
plotting. We should not be mutating original data, so replace `HACK`
with explicit check that `scalars` is not the original data and force
a copy.

Also, remove added `kwargs` to method since choosing whether or not to
scale data is likely not needed in most cases (and unrelated to the
original bug that spawned this PR).

* fix(revert): add `HACK` back

Iterating over `volume.array_names` in a `for` loops creates a
mutation. A list comprehension could be used, but instead these
mutations should be reviewed in a separate PR.

Adding the `HACK` back still allows the original issue of the PR
to be solved.

* Update pyvista/plotting/plotting.py

Co-authored-by: Hendry, Adam <adam.hendry@medtronic.com>
Co-authored-by: Alex Kaszynski <akascap@gmail.com>