Add support for NIFTI files with .nii and .nii.gz extensions.
NIFTI is a common format used in medical imaging extending the
Analyze™ 7.5 format:
https://nifti.nimh.nih.gov/nifti-1
In the most common form, NIFTI filenames have the .nii extension
or .hdr + .img when using the dual file format.
For compressed images, the convention is to use .nii.gz.

To read the files, we use the vtkNIFTIImageReader built into vtk.
This commit therefore also extends the get_ext function to support
the composed extensions with the .gz suffix.

The .hdr extension collides with the Radiance HDR file reader which is
already handled by vtkHDRReader. A way around is to force_ext='.nii'
with such files.