Addition of comment explaining 'indexing="ij"'
Koyama-san asked me to add an explanation about the changes in docstring.  (https://github.com/pyvista/pyvista/pull/4365#discussion_r1182029528)

I'm not sure where to put the explanation for this change in the docstring of function `voxelize`, because this particular change is about internal behavior of the function `voxelize`. So I added comment on line 72 and 73.  Possible position for this explanation might be line 13, before Parameters block of the docstring, but I'm afraid this explanation does not fit in docstring.

I'll rewrite these on further request. But, would you please review this version. 
Thank you.

x, y, z = np.meshgrid(x, y, z, indexing='ij')