Use relative paths in the offlineviewer directive
Absolute paths work only if the documentation is served from
the root of the domain. This is not the case for example when
serving multiple documentation versions from the same domain.

To get the relative path of the '_static' directory to the current
HTML document, we assume that the output structure is the same as
the source structure.